#67 rename class header become title in file html, rename function handleViewUser become handleEditUser in user folder controllers, rename file changeBaseImg become modules, remove file header scss and style for title
