+++ bigPractice/src/index.html
@@ -8,8 +8,8 @@
         <title>User Management</title>
     </head>
     <body>
-        <header class="header">
-            <h1 class="header__title">User Manager</h1>
+        <header class="title">
+            <h1 class="title__head">User Manager</h1>
         </header>
         <main class="container">
             <nav class="navbar">
