+++ bigPractice/src/styles/components/_index.scss
@@ -6,9 +6,9 @@
 @import "modal";
 @import "form";
 @import "icons";
-@import 'info';
-@import 'update';
-@import 'label';
-@import 'search';
-@import 'popup';
-
+@import "info";
+@import "update";
+@import "label";
+@import "search";
+@import "popup";
+@import "title";
