+++ bigPractice/src/styles/components/_title.scss
@@ -1,8 +1,8 @@
-.header {
+.title {
   width: 100%;
   background-color: $color-light;
 
-  &__title {
+  &__head {
     padding: 20px;
     font-size: $font-size-xl;
     color: $color-primary;
