+++ bigPractice/src/styles/layouts/_index.scss
@@ -1,3 +1,2 @@
 @import"body";
-@import "header";
 @import "container";
